Render the user's name, email address, profile image, and create/modify date as time-ago
Code and formatting suggested by Cursor. Initial implementation tried to use an <Image> React widget, but that presumed SSR (server-side rendering). Initial implementation also assumed the given URL could simply be embedded in the page. It seems that no longer works because of CORS. What does work is for the back-end to fetch the image, Base64 encode it, and include that data with the rest of the JSON. Then it can be directly interpreted by the browser (inline image data).