Refactor BaseEditable.tsx to drop the destructuring
I read up on this syntax and decided I just don't like the repetition. I'd rather the function accept an instance of BaseEditableProps, call it 'p', and then use syntax like 'p.value' to pull out that field, instead of having to repeat the whole list of field names just so we can write 'value' in the function body.