Rename useEditable to useFieldEditor
The name 'Editable' wasn't helping me to understand what this is for. 'FieldEditor' made a bit more sense.

I was mystified by the insistence on 'use' as a prefix, but I've learned that this is a React convention indicating a 'hook', and more than a convention, it tells React how to validate the code. So I'd better stick with it.