FieldEditorState interface to encapsulate the return values from useFieldEditor
I don't care for returning a list of seven values that must be assigned to variables in a particular order. In general I don't like positional parameter lists longer than 2 or 3 elements. Better to declare another interface, giving each returned value a key/value pair with a declared type. Clarity is improved on the calling side because it writes fieldEditor.x instead of x, so it's clear where the value comes from.