Fix display and handling of links
In EditableText components with isLink={true}, links now render in blue, and when the mouse hovers over them, they are undelined, so you can immediately tell they are links.

In MarkdownEditableText when a link is rendered, the same display issue is fixed. Also, when a link is clicked, it is rendered in a new tab/window, not the same tab.