Fix bug in POST user call
Forgot to move the user ID field to the end.

Also, I'm considering whether it's ever necessary to allow users to see each others' contact lists. This should probably be considered private information. Therefore it might never be necessary to accept a user ID in the calls to add/modify/delete contacts. Maybe an admin might need to do it. Maybe.