Start of admin UI for user search (by Cursor)
Cursor wrote a lot of this, but seemed to get stuck, possibly because of the requirement to kick off searches automatically after interaction with the input controls (e.g. entering text in the name field) followed by a period of inactivity. Right now it shows the search controls but has a "coming soon" in place of the results.